Add ability to specify metrics
So that the user can specify which metrics are computed when calling threshold_perf(), we add a new `metrics` argument that accepts a `yardstick` metric set.

By default this argument is `NULL`, in which case the function computes the default metrics. If not `NULL`, the function checks that it's an appropriate metric set and computes only the metrics provided in the set.